// src/app/models/employee/policy/location.ts
export class LocationModel {
    constructor() {
     
    }
    
    location_id: string = "0";
    location_code: string = "";
    location_name_th: string = "";
    location_name_en: string = "";
    location_detail: string = "";    
    modified_by: string = "";
    modified_date: string = "";
    index: number = 0;
    select: boolean = false;  
  }
  